Fix: update shiny filter for Heresy weapons
